Moves two function implementations and refactors the output buffer code to correctly handle byte-by-byte output translation.
[git-p4: depot-paths = "//depot/openwatcom/": change = 35054]

+++ bld/wgml/h/outbuff.h
@@ -45,7 +45,7 @@ extern "C" {    /* Use "C" linkage when in C++ mode. */
 #endif
 
 extern void ob_flush( void );
-extern void ob_insert_block( uint8_t * in_block, size_t count, bool out_trans, bool out_text, uint32_t font );
+extern void ob_insert_block( uint8_t * in_block, size_t count, bool out_trans, bool out_text, uint8_t font );
 extern void ob_insert_byte( uint8_t in_char );
 extern void ob_insert_ps_text_start( void );
 extern void ob_setup( void );
